+++ database/wizard_utils.go
@@ -24,7 +24,7 @@ func (d *GormDatabase) WizardNewNetworkDevicePoint(plugin string, network *model
 	}
 
 	network.PluginConfId = p.UUID
-	n, err := d.CreateNetwork(network, false)
+	n, err := d.CreateNetwork(network)
 	if err != nil {
 		return nil, fmt.Errorf("network creation failure: %s", err)
 	}
@@ -39,7 +39,7 @@ func (d *GormDatabase) WizardNewNetworkDevicePoint(plugin string, network *model
 
 	if point != nil {
 		point.DeviceUUID = dev.UUID
-		_, err = d.CreatePoint(point, false)
+		_, err = d.CreatePoint(point)
 		if err != nil {
 			return nil, fmt.Errorf("consumer point creation failure: %s", err)
 		}