+++ plugin/nube/protocals/bacnetmaster/app.go
@@ -23,7 +23,7 @@ func (inst *Instance) addNetwork(body *model.Network) (network *model.Network, e
 	}
 
 	inst.bacnetDebugMsg("addNetwork(): ", body.Name)
-	network, err = inst.db.CreateNetwork(body, true)
+	network, err = inst.db.CreateNetwork(body)
 	if network == nil || err != nil {
 		inst.bacnetErrorMsg("addNetwork(): failed to create bacnet network: ", body.Name)
 		if err != nil {
@@ -122,7 +122,7 @@ func (inst *Instance) addPoint(body *model.Point) (point *model.Point, err error
 	isOutput := checkForOutputType(body.ObjectType)
 	body.IsOutput = nils.NewBool(isOutput)
 
-	point, err = inst.db.CreatePoint(body, true, true)
+	point, err = inst.db.CreatePoint(body, true)
 	if point == nil || err != nil {
 		inst.bacnetDebugMsg("addPoint(): failed to create bacnet point: ", body.Name)
 		return nil, err
@@ -175,7 +175,7 @@ func (inst *Instance) updateNetwork(body *model.Network) (network *model.Network
 		body.CommonFault.LastOk = time.Now().UTC()
 	}
 
-	network, err = inst.db.UpdateNetwork(body.UUID, body, true)
+	network, err = inst.db.UpdateNetwork(body.UUID, body)
 	if err != nil || network == nil {
 		return nil, err
 	}
@@ -208,7 +208,7 @@ func (inst *Instance) updateNetwork(body *model.Network) (network *model.Network
 		inst.db.ClearErrorsForAllDevicesOnNetwork(network.UUID, true)
 	}
 
-	network, err = inst.db.UpdateNetwork(body.UUID, network, true)
+	network, err = inst.db.UpdateNetwork(body.UUID, network)
 	if err != nil || network == nil {
 		return nil, err
 	}
@@ -236,7 +236,7 @@ func (inst *Instance) updateDevice(body *model.Device) (device *model.Device, er
 		body.CommonFault.LastOk = time.Now().UTC()
 	}
 
-	device, err = inst.db.UpdateDevice(body.UUID, body, true)
+	device, err = inst.db.UpdateDevice(body.UUID, body)
 	if err != nil {
 		return nil, err
 	}
@@ -297,7 +297,7 @@ func (inst *Instance) updateDevice(body *model.Device) (device *model.Device, er
 	}
 	// TODO: NEED TO ACCOUNT FOR OTHER CHANGES ON DEVICE.  It would be useful to have a way to know if the device polling rates were changed.
 
-	device, err = inst.db.UpdateDevice(device.UUID, body, true)
+	device, err = inst.db.UpdateDevice(device.UUID, body)
 	if err != nil {
 		return nil, err
 	}
@@ -335,11 +335,11 @@ func (inst *Instance) updatePoint(body *model.Point) (point *model.Point, err er
 		body.CommonFault.Message = "point disabled"
 		body.CommonFault.LastFail = time.Now().UTC()
 	}
-	point.CommonFault.InFault = false
-	point.CommonFault.MessageLevel = model.MessageLevel.Info
-	point.CommonFault.MessageCode = model.CommonFaultCode.PointWriteOk
-	point.CommonFault.Message = fmt.Sprintf("last-updated: %s", utilstime.TimeStamp())
-	point.CommonFault.LastOk = time.Now().UTC()
+	body.CommonFault.InFault = false
+	body.CommonFault.MessageLevel = model.MessageLevel.Info
+	body.CommonFault.MessageCode = model.CommonFaultCode.PointWriteOk
+	body.CommonFault.Message = fmt.Sprintf("last-updated: %s", utilstime.TimeStamp())
+	body.CommonFault.LastOk = time.Now().UTC()
 	point, err = inst.db.UpdatePoint(body.UUID, body)
 	if err != nil || point == nil {
 		inst.bacnetDebugMsg("updatePoint(): bad response from UpdatePoint() err:", err)