+++ plugin/nube/protocals/modbus/app.go
@@ -19,7 +19,7 @@ import (
 // THE FOLLOWING GROUP OF FUNCTIONS ARE THE PLUGIN RESPONSES TO API CALLS FOR PLUGIN POINT, DEVICE, NETWORK (CRUD)
 func (inst *Instance) addNetwork(body *model.Network) (network *model.Network, err error) {
 	inst.modbusDebugMsg("addNetwork(): ", body.Name)
-	network, err = inst.db.CreateNetwork(body, true)
+	network, err = inst.db.CreateNetwork(body)
 	if err != nil {
 		inst.modbusErrorMsg("addNetwork(): failed to create modbus network: ", body.Name)
 		return nil, errors.New("failed to create modbus network")
@@ -82,7 +82,7 @@ func (inst *Instance) addPoint(body *model.Point) (point *model.Point, err error
 	isOutput := checkForOutputType(body.ObjectType)
 	body.IsOutput = nils.NewBool(isOutput)
 
-	point, err = inst.db.CreatePoint(body, true, true)
+	point, err = inst.db.CreatePoint(body, true)
 	if point == nil || err != nil {
 		inst.modbusDebugMsg("addPoint(): failed to create modbus point: ", body.Name)
 		return nil, errors.New(fmt.Sprint("failed to create modbus point. err: ", err))
@@ -136,7 +136,7 @@ func (inst *Instance) updateNetwork(body *model.Network) (network *model.Network
 		body.CommonFault.LastOk = time.Now().UTC()
 	}
 
-	network, err = inst.db.UpdateNetwork(body.UUID, body, true)
+	network, err = inst.db.UpdateNetwork(body.UUID, body)
 	if err != nil || network == nil {
 		return nil, err
 	}
@@ -165,7 +165,7 @@ func (inst *Instance) updateNetwork(body *model.Network) (network *model.Network
 		inst.db.ClearErrorsForAllDevicesOnNetwork(network.UUID, true)
 	}
 
-	network, err = inst.db.UpdateNetwork(body.UUID, network, true)
+	network, err = inst.db.UpdateNetwork(body.UUID, network)
 	if err != nil || network == nil {
 		return nil, err
 	}
@@ -193,7 +193,7 @@ func (inst *Instance) updateDevice(body *model.Device) (device *model.Device, er
 		body.CommonFault.LastOk = time.Now().UTC()
 	}
 
-	device, err = inst.db.UpdateDevice(body.UUID, body, true)
+	device, err = inst.db.UpdateDevice(body.UUID, body)
 	if err != nil || device == nil {
 		return nil, err
 	}
@@ -247,7 +247,7 @@ func (inst *Instance) updateDevice(body *model.Device) (device *model.Device, er
 	}
 	// TODO: NEED TO ACCOUNT FOR OTHER CHANGES ON DEVICE.  It would be useful to have a way to know if the device polling rates were changed.
 
-	device, err = inst.db.UpdateDevice(device.UUID, device, true)
+	device, err = inst.db.UpdateDevice(device.UUID, device)
 	if err != nil {
 		return nil, err
 	}
@@ -285,12 +285,12 @@ func (inst *Instance) updatePoint(body *model.Point) (point *model.Point, err er
 		body.CommonFault.Message = "point disabled"
 		body.CommonFault.LastFail = time.Now().UTC()
 	}
-	point.CommonFault.InFault = false
-	point.CommonFault.MessageLevel = model.MessageLevel.Info
-	point.CommonFault.MessageCode = model.CommonFaultCode.PointWriteOk
-	point.CommonFault.Message = fmt.Sprintf("last-updated: %s", utilstime.TimeStamp())
-	point.CommonFault.LastOk = time.Now().UTC()
-	point, err = inst.db.UpdatePoint(body.UUID, body)
+	body.CommonFault.InFault = false
+	body.CommonFault.MessageLevel = model.MessageLevel.Info
+	body.CommonFault.MessageCode = model.CommonFaultCode.PointWriteOk
+	body.CommonFault.Message = fmt.Sprintf("last-updated: %s", utilstime.TimeStamp())
+	body.CommonFault.LastOk = time.Now().UTC()
+	body, err = inst.db.UpdatePoint(body.UUID, body)
 	if err != nil || point == nil {
 		inst.modbusDebugMsg("updatePoint(): bad response from UpdatePoint() err:", err)
 		return nil, err